refactor: gimie.sources -> gimie.extractors
