+++ gimie/extractors/common/queries.py
